Adb: Reset device in initialization
webadb.js fails to connect to my device on Linux:
A transfer error has occurred.

However, with the same configurations, it works on Windows. This
indicates handling of USB devices of host (AMD X570) on Linux
might lead to undesired results. Though, without further testing,
the interference from userspace (Ubuntu 20.04, KDE) can not be
ruled out.

Thus, this patch calls reset() to ensure the USB port is in a
clean state before we send the data.

Signed-off-by: Jesse Chan <jc@linux.com>